+++ client/Routes.js
@@ -8,6 +8,7 @@ import SingleProduct from './components/SingleProduct'
 import Cart from './components/Cart';
 import Login from './components/Login'
 import Signup from './components/Signup'
+import Orders from './components/Orders';
 
 /**
  * COMPONENT
@@ -28,6 +29,7 @@ class Routes extends Component {
             <Route exact path="/products" component={AllProducts} />
             <Route path="/products/:id" component={SingleProduct} />
             <Route exact path="/cart" component={Cart} />
+            <Route exact path="/orders" component={Orders} />
             <Redirect to="/home" />
           </Switch>
         ) : (
