+++ client/components/Navbar.js
@@ -14,6 +14,7 @@ const Navbar = ({ handleClick, isLoggedIn }) => (
           <Link to="/home">Home</Link>
           <Link to="/products">Products</Link>
           <Link to="/cart">Cart</Link>
+          <Link to="/orders">Order History</Link>
           <a href="#" onClick={handleClick}>
             Logout
           </a>
